refactor(xtask): separate gates first-failure parsing module (#0000)
